显示股票的涨跌颜色，涨为红色，跌为绿色

+++ Plugins/Stock/Stock.cpp
@@ -118,7 +118,7 @@ void Stock::ParseJsonData(std::string json_data)
 
         StockInfo& StockInfo = g_data.GetStockInfo(CCommon::StrToUnicode(key.c_str()));
 
-        int data_size = data_arr.size();
+        size_t data_size = data_arr.size();
 
         CString name;
         float now = -1;
