+++ utilities/Common.h
@@ -30,10 +30,10 @@ namespace utilities
         {
             if (str.empty()) return;
 
-            int size = str.size();  //字符串的长度
+            size_t size = str.size();  //字符串的长度
             if (size < 0) return;
-            int index1 = 0;     //字符串中第1个不是空格或控制字符的位置
-            int index2 = size - 1;  //字符串中最后一个不是空格或控制字符的位置
+            size_t index1 = 0;     //字符串中第1个不是空格或控制字符的位置
+            size_t index2 = size - 1;  //字符串中最后一个不是空格或控制字符的位置
             while (index1 < size && str[index1] >= 0 && str[index1] <= 32)
                 index1++;
             while (index2 >= 0 && str[index2] >= 0 && str[index2] <= 32)
